Move standing policy review to the fall
The spring is pretty busy for the EC (or at least the P/VP) due to needing to
organize elections. Fall is a little less busy, so move standing policy review
then. (This also avoids risk of a "lame duck" renewal process if it happens mid
to late in term, rather than early.)