Further tweak to appointed officers
Ginda: "I would drop the word "new" from 2.C (or replace it with "the"), unless
the expectation is that the appointed officers will all be new each year. That
WAS the rule for appointed members-at-large, fwiw, and the wording was similar.
And I don't think it would be healthy for the club to actually turn over all
the appointed officers every year. I think some level of stability and
institutional knowledge is valuable, and the standing policy would still
clearly say that the EC needs to review all the positions."