Fix weird \url linebreak in HTML
I'd thought \linebreak was a hint (the summary at the top of
http://www.personal.ceu.hu/tex/breaking.htm is misleading) and figured surely
HTML would ignore such hints, but it looks like it's actually a demand and the
HTML respects it. We now use \linebreak only for the LaTeX build and leave it
out for the HTML one, which solves this problem.