Build using an Ubuntu image
latex2html isn't preinstalled on the LaTeX image, so it's not actually clear
using an official image will be slower (especially if it's cached), and I'd
feel slightly more comfortable using one with a more known creator. Also, when
I *did* install latex2html on the old image, it mysteriously failed. This works
much better.